boss Event script update
The script for the boss event is now working. Lasers are removed prior to the player entering the boss area. Lasers then appear and a collider is activated so that the player cannot leave until the boss is dead. Once the boss dies. Lasers disappear once more and the collider is deactivated.